% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_coordinates.R
\name{check_coordinates}
\alias{check_coordinates}
\title{Checks if the coordinates are in Central Park}
\usage{
check_coordinates(latitude, longitude)
}
\arguments{
\item{latitude}{numeric. A vector with the latitude.}

\item{longitude}{numeric. A vector with the longitude.}
}
\value{
a vector of logicals
}
\description{
Checks if the coordinates are in Central Park
}
\examples{
check_coordinates(latitude = 42, longitude = 0)
check_coordinates(latitude = 40.8, longitude = -73.950)
}
