% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_plus_in_primary_fur_color.R
\name{check_plus_in_primary_fur_color}
\alias{check_plus_in_primary_fur_color}
\title{Check the values of primary fur color}
\usage{
check_plus_in_primary_fur_color(primary_fur_color)
}
\arguments{
\item{primary_fur_color}{Character. A vector with the primary fur color.}
}
\value{
a vector of logicals
}
\description{
Check the values of primary fur color
}
\examples{
check_plus_in_primary_fur_color(primary_fur_color = c("Cinnamon", "Blue", "Yellow"))
check_plus_in_primary_fur_color(primary_fur_color = c("Cinnamon", "Blue", NA))
check_plus_in_primary_fur_color(primary_fur_color = c("Cinnamon", "Blue", 3))
check_plus_in_primary_fur_color(primary_fur_color = c("Ciinamon+blue", "yellow"))
}
