% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_df_nrows.R
\name{get_df_nrows}
\alias{get_df_nrows}
\title{get nrows of input df}
\usage{
get_df_nrows(df)
}
\arguments{
\item{df}{a datframe for which one wants to know the number of rows}
}
\value{
a message indicating the number of rows hold by the df
}
\description{
get nrows of input df
}
\examples{
get_df_nrows(iris)
get_df_nrows(mtcars)
}
